% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropas-package.R
\docType{package}
\name{ropas-package}
\alias{ropas}
\alias{ropas-package}
\title{ropas: R Client for OPAS Environmental Monitoring Web Services}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides an R interface to the OPAS (OPen Air System) REST API exposed by ISPRA (Istituto Superiore per la Protezione e la Ricerca Ambientale). Functions cover authentication (JWT login and token refresh), retrieval of the national air quality monitoring network catalogue (stations, parameters, data series), and download of hourly and daily time series measurements with validity codes.
}
\author{
\strong{Maintainer}: Giovanni Bonafe \email{giovanni.bonafe@arpa.fvg.it} (\href{https://orcid.org/0000-0002-7815-8866}{ORCID})

Authors:
\itemize{
  \item Giovanni Bonafe \email{giovanni.bonafe@arpa.fvg.it} (\href{https://orcid.org/0000-0002-7815-8866}{ORCID})
}

}
\keyword{internal}
